requirements.txt
apache-airflow==2.7.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
pytest==7.4.0
pytest-mock==3.11.1
google-api-python-client==2.108.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydrive2==1.15.4
